Animating the player via code
In this section we have learned how to animate the player via code. movementX is much helpful as GetAxisRaw will only generates three values which is -1,0,1. So using this, if the player moves left or right we are animating respectively. FlipX is a default component present in spriterenderer which is used to flip the entire player in which direction the player in moving.